requirements.txt
requests
pandas
psycopg2-binary
yfinance
numpy
tqdm
beautifulsoup4
lxml
ollama
fastapi
uvicorn[standard]
apscheduler
SQLAlchemy
pdfplumber
trafilatura